% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapfunc.R
\name{auc}
\alias{auc}
\title{Compute AUC using a GWAS summary statistic dataset for validation
'\code{get.correlation.adj} computes a summation of the genotypic correlation between PGS model and a reference panel}
\usage{
auc(
  prs.model.file,
  gwas.summary.stats.file,
  N0,
  N1,
  soFile = "getAdjCorrelation.so",
  KG.plink.pre = "KG.all.chr",
  pos_thr = 5e+08,
  flag.correlation.adj.inputted.data = FALSE
)
}
\arguments{
\item{prs.model.file}{a data.table containing the PGS model, ideally an output from \code{computePGS}}

\item{gwas.summary.stats.file}{a vector representing the name of the GWAS sumstats validation dataset}

\item{N0}{a scalar. Number of controls in the GWAS validation dataset}

\item{N1}{a scalar. Number of cases in the GWAS validation dataset}

\item{soFile}{a string of the path to the compiled .so file for computation, provided by SummaryAUC}

\item{pos_thr}{a scalar representing the threshold for distance of SNPs, correlation will be caculated for any SNP pairs within this distance}

\item{flag.correction.ajd.inputted.data}{a logical. Are there inputted SNPs in the GWAS validation dataset? (Not tested in our implementation)}
}
\value{
a list containing AUC, AUC variance, adj.cor, and number of SNPs in common in the PRS model and the validaton dataset
}
\description{
This function, adapted from SummaryAUC (https://github.com/lsncibb/SummaryAUC)
source code, will use a PGS model of a binary trait and a GWAS summary statistics dataset for validation.
Ideally, PGS model will have a specific format, containing 3 columns: pid (chr:pos in hg19 build), ALT, and weights.
GWAS validation dataset will contain: pid, ALT, MAF, BETA, and P.
}
\details{
For more details on the method, please see original publication (Song et al., 2019; https://academic.oup.com/bioinformatics/article/35/20/4038/5419857).
}
