% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EUR_ld.blocks38}
\alias{EUR_ld.blocks38}
\title{LD block architecture for European populations (hg38).}
\format{
A GRanges object containing 1361 ranges
\describe{
\item{seqnames}{chromosome}
\item{ranges}{start and stop positions for the block}
\item{strand}{genomic strand, irrelevant here}
}
}
\source{
\url{https://github.com/jmacdon/LDblocks_GRCh38/tree/master/data}
}
\usage{
EUR_ld.blocks38
}
\description{
A GRanges object containing the LD block for European ancestry, in hg38 build.
This dataset was obtained from \href{https://www.biorxiv.org/content/10.1101/2022.03.04.483057v2}{MacDonald et al. (2022)}, in bed format,
then transformed to GRanges.
See manuscript for more details.
}
\keyword{datasets}
