% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapfunc.R
\name{find_file_in_ftp}
\alias{find_file_in_ftp}
\title{Finding a file in an FTP directory
This is an internal function to help \code{gwascat.download} find the right file}
\usage{
find_file_in_ftp(ftp_address, acc, hm)
}
\arguments{
\item{ftp_address}{a string. An FTP address provided by \code{gwascat.download}.}

\item{acc}{a string containing the accession for the desired study.}

\item{hm}{a logical. Should it look in the harmonised directory?}
}
\value{
a data.table containing the dataset.
}
\description{
Finding a file in an FTP directory
This is an internal function to help \code{gwascat.download} find the right file
}
\author{
Guillermo Reales
}
