% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapfunc.R
\name{g.class}
\alias{g.class}
\title{g.class}
\usage{
g.class(x, y)
}
\arguments{
\item{x}{vector of allele codes from dataset X}

\item{y}{vector of allele codes from dataset Y, same length as x}
}
\value{
character vector of allele switching classes
}
\description{
Define possible allele switching classes
}
\details{
This function is verbatim of its namesake in annotSnpStats package (github.com/chr1swallace/annotSnpStats/), by Chris Wallace
}
\examples{
alleles.X <- c(snp1="A/G",snp2="A/G",snp3="A/G",snp4="A/G",snp5="A/T",snp6="A/T")
alleles.Y <- c(snp1="A/G",snp2="G/A",snp3="T/C",snp4="C/T",snp5="A/T",snp6="T/A")
classes <- g.class(x=alleles.X,y=alleles.Y)
cbind(alleles.X,alleles.Y,classes)
}
\author{
Chris Wallace
}
