% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapfunc.R
\name{g.complement}
\alias{g.complement}
\title{g.complement}
\usage{
g.complement(x)
}
\arguments{
\item{x}{character vector of genotypes}
}
\value{
character vector of genotypes on the alternative strand
}
\description{
Complement genotypes
}
\details{
ie A/G -> T/C
This function was copycat from annotSnpStats package (github.com/chr1swallace/annotSnpStats/), by Chris Wallace
}
\examples{
g.complement(c("A/G","A/T"))
}
\author{
Chris Wallace
}
