% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapfunc.R
\name{g.rev}
\alias{g.rev}
\title{g.rev}
\usage{
g.rev(x, sep = "/")
}
\arguments{
\item{x}{character vector of genotypes}

\item{sep}{character with which to separate alleles. Default is "/".}
}
\value{
character vector of reversed genotypes
}
\description{
Reverse alleles in a genotype
}
\details{
ie A/G -> G/A
This function is verbatim of its namesake from annotSnpStats package (github.com/chr1swallace/annotSnpStats/), by Chris Wallace
}
\examples{
g.rev(c("A/G","A/T"))
}
\author{
Chris Wallace
}
