% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapfunc.R
\name{get.correlation.adj}
\alias{get.correlation.adj}
\title{Get SNP correlation summation
'\code{get.correlation.adj} computes a summation of the genotypic correlation between PGS model and a reference panel}
\usage{
get.correlation.adj(
  prs.model,
  tau,
  KG.plink.pre = "KG.all.chr",
  soFile = "scripts/getAdjCorrelation.so",
  pos_thr = 5e+08
)
}
\arguments{
\item{prs.model}{a data.table containing the PGS model, ideally an output from \code{computePGS}}

\item{tau}{a numeric vector representing the genotypic variance in the GWAS validation dataset (sqrt(2 * MAF * (1 - MAF) * INFO)}

\item{KG.plink.pre}{a string representing the prefix of the reference plink files, provided by SummaryAUC}

\item{soFile}{a string of the path to the compiled .so file for computation, provided by SummaryAUC}

\item{pos_thr}{a scalar representing the threshold for distance of SNPs, correlation will be caculated for any SNP pairs within this distance}
}
\value{
a scalar. A summation of beta * tau * cor for SNPs
}
\description{
This function, taken from SummaryAUC (https://github.com/lsncibb/SummaryAUC)
source code, will use a reference panel from 1000 Genomes Project (379 individuals
of European ancestry) to calculate the genotypic correlation between SNPs in the PGS
model and the panel, outputting a summation that will be later use at the denominator
for delta calculation in \code{auc} computation.
For more details, please see original publication (Song et al., 2019; https://academic.oup.com/bioinformatics/article/35/20/4038/5419857).
}
