% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapfunc.R
\name{logsum}
\alias{logsum}
\title{Helper function to sum logs without loss of precision}
\usage{
logsum(x)
}
\arguments{
\item{x}{a vector of logs to sum}
}
\value{
a scalar
}
\description{
Sums logs without loss of precision
This function is verbatim of its namesake in cupcake package (github.com/ollyburren/cupcake/)
}
\author{
Chris Wallace
}
