% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{michailidou}
\alias{michailidou}
\title{Subset of Michailidou BRCA GWAS sumstat dataset.}
\format{
A data.table object containing 100,000 SNPs
\describe{
\item{hm_rsid}{rsids, or SNP ids}
\item{hm_chrom}{chromosome}
\item{hm_pos}{base position, in hg38}
\item{hm_other_allele}{reference, or non-effect allele}
\item{hm_effect_allele}{alternative, or effect allele}
\item{hm_beta}{beta, log(OR), or effect size}
\item{hm_effect_allele_frequency}{effect allele frequency}
\item{standard_error}{standard error of beta}
\item{p_value}{p-value}
}
}
\source{
\url{ftp://ftp.ebi.ac.uk/pub/databases/gwas/summary_statistics/MichailidouK_29059683_GCST004988/harmonised/29059683-GCST004988-EFO_0000305.h.tsv.gz}
}
\usage{
michailidou
}
\description{
A data.table containing a subset of \href{https://www.nature.com/articles/nature24284/}{Michailidou et al., 2017} breast cancer summary statistic dataset, in hg38 build.
This dataset is freely available in GWAS catalog (see link below). I removed unnecessary and all-missing columns, and rows
with missing data at hm_beta and hm_effect_allele_frequency, and took a random sample of 100,000 SNPs without replacement.
}
\keyword{datasets}
