% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{michailidou19}
\alias{michailidou19}
\title{Subset of Michailidou BRCA GWAS sumstat dataset.}
\format{
A data.table object containing 100,000 SNPs
\describe{
SNPID, CHR, BP, REF, ALT, ALT_FREQ, BETA, SE, P
\item{SNPID}{rsids, or SNP ids}
\item{CHR}{chromosome}
\item{BP}{base position, in hg38}
\item{REF}{reference, or non-effect allele}
\item{ALT}{alternative, or effect allele}
\item{ALT_FREQ}{effect allele frequency}
\item{BETA}{beta, log(OR), or effect size}
\item{SE}{standard error of beta}
\item{P}{p-value}
}
}
\source{
\url{ftp://ftp.ebi.ac.uk/pub/databases/gwas/summary_statistics/MichailidouK_29059683_GCST004988/harmonised/29059683-GCST004988-EFO_0000305.h.tsv.gz}
}
\usage{
michailidou19
}
\description{
A data.table containing a subset of \href{https://www.nature.com/articles/nature24284/}{Michailidou et al., 2017} breast cancer summary statistic dataset, in hg19 build.
This dataset is freely available in GWAS catalog (see link below). I used "chromosome", "base_pair_location columns", removed unnecessary and all-missing columns, and took a random sample of 100,000 SNPs without replacement.
}
\keyword{datasets}
