% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EUR_ld.blocks38}
\alias{EUR_ld.blocks38}
\title{LD block architecture for European populations (hg38).}
\format{
A GRanges object containing 1625 ranges
\describe{
\item{seqnames}{chromosome}
\item{ranges}{start and stop positions for the block}
\item{strand}{genomic strand, irrelevant here}
}
}
\source{
\url{https://bitbucket.org/nygcresearch/ldetect-data/src}
}
\usage{
EUR_ld.blocks38
}
\description{
A GRanges object containing the LD block for European ancestry, in hg38 build.
This dataset was obtained from \href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4731402/}{Belisa and Pickrell (2016)}, in bed format,
then liftovered to hg38 using UCSC liftOver tool, then converted to GRanges.
See manuscript for more details.
}
\keyword{datasets}
