% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapfunc.R
\name{pgs.file.preprocess}
\alias{pgs.file.preprocess}
\title{Preprocess GWAS summary statistics datasets for PGS computation}
\usage{
pgs.file.preprocess(dataset, blockfile = "ld.blocks.RDS", ref = NULL)
}
\arguments{
\item{dataset}{a string with the name or path to the GWAS summary statistic dataset, or the path to it}

\item{blockfile}{a string with the name or path to a RDS file containing the genomic coordinates of LD blocks, computed for EUR populations.}

\item{ref}{a string indicating the reference file SNPs should be filtered and aligned to.}
}
\value{
a data.table to be used for PGS computation.
}
\description{
\code{pgs.file.preprocess} imports the dataset, checks for minimum information, assign LD blocks to SNPs from an externally calculated LD architecture,
and optionally filters SNPs and aligns alleles to the reference HapMap3 panel.
}
\details{
The GWAS summary statistics file to compute PGS using our method must contain the following minimum columns, with these exact column names:
CHR19: Chromosome
BP19: Base position (in GRCh37/hg19)
REF: Reference, or non-effect allele
ALT: Alternative, or effect allele, the one \eqn{\beta} refers to
BETA: Beta (or log(OR)), or effect sizes.
SE: Standard error of \eqn{\beta}
P: P-values for the association test
ALT_FREQ: Minor/ALT allele frequency in the tested population, or in a close population from a reference panel. No problem it flipped.

If a ref is provided. It should have 4 columns: CHR19, BP19, SNPID, REF, and ALT
}
\author{
Guillermo Reales, Chris Wallace
}
